requirements_macosM1.txt
absl-py==1.4.0
appdirs @ file:///home/conda/feedstock_root/build_artifacts/appdirs_1603108395799/work
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1649077682618/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
astunparse==1.6.3
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1666764727380/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools==5.3.0
certifi==2022.12.7
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1671179605388/work
charset-normalizer==3.0.1
click @ file:///home/conda/feedstock_root/build_artifacts/click_1666798198223/work
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1674202310934/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography-split_1672673470954/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
Cython @ file:///Users/runner/miniforge3/conda-bld/cython_1673054151758/work
cytoolz @ file:///Users/runner/miniforge3/conda-bld/cytoolz_1666829931211/work
dask @ file:///home/conda/feedstock_root/build_artifacts/dask-core_1674873905639/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
flatbuffers==23.1.21
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1666827220469/work
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1674184942191/work
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py @ file:///Users/runner/miniforge3/conda-bld/h5py_1674498930724/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imagecodecs @ file:///Users/runner/miniforge3/conda-bld/imagecodecs_1674483163412/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1674460904928/work
importlib-metadata==6.0.0
ipdb @ file:///home/conda/feedstock_root/build_artifacts/ipdb_1671923457576/work
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1674912041953/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1663332044897/work
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1666805848360/work
libclang==15.0.6.1
locket @ file:///home/conda/feedstock_root/build_artifacts/locket_1650660393415/work
lxml @ file:///Users/runner/miniforge3/conda-bld/lxml_1671013608894/work
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-suite_1637205352958/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
munkres==1.1.4
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1673151334029/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1675039124141/work
oauthlib==3.2.2
opencv-python==4.7.0
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1673482170163/work
pandas==1.4.4
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
partd @ file:///home/conda/feedstock_root/build_artifacts/partd_1660316728562/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1672675381886/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1643032624649/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1670414775770/work
protobuf==4.21.12
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydot @ file:///Users/runner/miniforge3/conda-bld/pydot_1666825761109/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1672682006896/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1672659226110/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1673864280276/work
PyWavelets @ file:///Users/runner/miniforge3/conda-bld/pywavelets_1673082401837/work
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1666772523114/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1673863902341/work
requests-oauthlib==1.3.1
rsa==4.9
scikit-image @ file:///Users/runner/miniforge3/conda-bld/scikit-image_1667117212579/work
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1674599054071/work
scipy==1.10.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
tensorflow-metal==0.6.0
termcolor==2.2.0
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1643647933166/work
tifffile @ file:///home/conda/feedstock_root/build_artifacts/tifffile_1674498532372/work
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1604308577558/work
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1657485559105/work
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1666788903714/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1662214488106/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1667240193675/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1673452138552/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
